+++ jeap-governance-domain/src/main/java/ch/admin/bit/jeap/governance/domain/rule/GracePeriodComponentEntry.java
@@ -1,0 +1,10 @@
+package ch.admin.bit.jeap.governance.domain.rule;
+
+import java.time.ZonedDateTime;
+
+public interface GracePeriodComponentEntry {
+    Long getSystemId();
+    Long getSystemComponentId();
+    String getRuleId();
+    ZonedDateTime getViolationDetectedAt();
+}