+++ jeap-governance-reporting/src/main/java/ch/admin/bit/jeap/governance/reporting/confluence/model/RuleReportComponentModel.java
@@ -14,4 +14,6 @@ public class RuleReportComponentModel {
     private String systemName;
     private String systemPageSuffix;
     private ZonedDateTime nonCompliantSince;
+    private ZonedDateTime violationDetectedAt;
+    private ZonedDateTime gracePeriodEndsAt;
 }